backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pypdf
python-docx
python-multipart
